Rename pv_slicer -> path_slicer; add shared modes and helpers
